GTK: Workaround ABI issue with Windows GTK2 bundle and GCC > 3
GtkScrolledWindow contains a bitfield, and GCC 3.4 and 4.8 don't agree
on the size of the structure (regardless of -mms-bitfields):

- GCC 3.4 has sizeof(GtkScrolledWindow)=88
- GCC 4.8 has sizeof(GtkScrolledWindow)=84

As Windows GTK2 bundle is built with GCC 3, it requires types derived
from GtkScrolledWindow to be at least 88 bytes, which means we need to
add some fake padding to fill in the extra 4 bytes.
There is however no other issue with the layout difference as we never
access any GtkScrolledWindow fields ourselves.

See http://lists.geany.org/pipermail/devel/2015-April/thread.html#9379